refactor(ui): split config view panels
